Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-grid/parsing/grid-area-valid-expected.txt


FAIL e.style['grid-area'] = "auto" should set the property value assert_equals: serialization should be canonical expected "auto / auto / auto / auto" but got "auto auto auto auto"
FAIL e.style['grid-area'] = "AuTo" should set the property value assert_equals: serialization should be canonical expected "auto / auto / auto / auto" but got "auto auto auto auto"
FAIL e.style['grid-row'] = "auto" should set the property value assert_equals: serialization should be canonical expected "auto / auto" but got "auto auto"
PASS e.style['grid-column-end'] = "AuTo" should set the property value 
FAIL e.style['grid-area'] = "--a" should set the property value assert_equals: serialization should be canonical expected "--a / --a / --a / --a" but got "\"--a\" \"--a\" \"--a\" \"--a\""
FAIL e.style['grid-row'] = "-zπ" should set the property value assert_equals: serialization should be canonical expected "-zπ / -zπ" but got "-zπ -zπ"
PASS e.style['grid-row-start'] = "AZ" should set the property value 
PASS e.style['grid-column-start'] = "-_π" should set the property value 
PASS e.style['grid-row-end'] = "_9" should set the property value 
FAIL e.style['grid-area'] = "1" should set the property value assert_equals: serialization should be canonical expected "1 / auto / auto / auto" but got "1 auto auto auto"
FAIL e.style['grid-area'] = "+90 -a-" should set the property value assert_equals: serialization should be canonical expected "90 -a- / auto / auto / auto" but got "90 -a- auto auto auto"
FAIL e.style['grid-row'] = "az 2" should set the property value assert_equals: serialization should be canonical expected "2 az / auto" but got "2 az auto"
FAIL e.style['grid-column'] = "9" should set the property value assert_equals: serialization should be canonical expected "9 / auto" but got "9 auto"
FAIL e.style['grid-column'] = "-19 zA" should set the property value assert_equals: serialization should be canonical expected "-19 zA / auto" but got "-19 zA auto"
FAIL e.style['grid-column'] = "-A0 33" should set the property value assert_equals: serialization should be canonical expected "33 -A0 / auto" but got "33 -A0 auto"
PASS e.style['grid-row-start'] = "-19" should set the property value 
PASS e.style['grid-row-start'] = "9 -Z_" should set the property value 
PASS e.style['grid-column-start'] = "+90" should set the property value 
PASS e.style['grid-column-start'] = "Z -44" should set the property value 
PASS e.style['grid-row-end'] = "1 -πA" should set the property value 
PASS e.style['grid-column-end'] = "π_ +5" should set the property value 
FAIL e.style['grid-area'] = "span 2 i" should set the property value assert_equals: serialization should be canonical expected "span 2 i / auto / auto / auto" but got "span 2 i auto auto auto"
FAIL e.style['grid-area'] = "i 2 SpAn" should set the property value assert_equals: serialization should be canonical expected "span 2 i / auto / auto / auto" but got "span 2 i auto auto auto"
FAIL e.style['grid-row'] = "span 2" should set the property value assert_equals: serialization should be canonical expected "span 2 / auto" but got "span 2 auto"
FAIL e.style['grid-column'] = "i SpAn" should set the property value assert_equals: serialization should be canonical expected "span i / auto" but got "span i auto"
PASS e.style['grid-row-start'] = "span i" should set the property value 
PASS e.style['grid-column-start'] = "SpAn i 2" should set the property value 
PASS e.style['grid-row-end'] = "2 i span" should set the property value 
PASS e.style['grid-column-end'] = "2 SpAn" should set the property value 
FAIL e.style['grid-area'] = "auto / i" should set the property value assert_equals: serialization should be canonical expected "auto / i / auto / i" but got "auto i auto i"
FAIL e.style['grid-area'] = "auto / i / 2 j" should set the property value assert_equals: serialization should be canonical expected "auto / i / 2 j / i" but got "auto i 2 j i"
FAIL e.style['grid-area'] = "auto / i / 2 j / span 3 k" should set the property value assert_equals: serialization should be canonical expected "auto / i / 2 j / span 3 k" but got "auto i 2 j span 3 k"
FAIL e.style['grid-row'] = "auto / i" should set the property value assert_equals: serialization should be canonical expected "auto / i" but got "auto i"
FAIL e.style['grid-column'] = "2 j / span 3 k" should set the property value assert_equals: serialization should be canonical expected "2 j / span 3 k" but got "2 j span 3 k"
PASS e.style['grid-column-end'] = "\\31st" should set the property value 
PASS e.style['grid-column-end'] = "\\31 st" should set the property value 
